meson-omit-meson-dist.patch: fix meson.build mess
meson.build requres build-aux/meson-dist file which is not
present in the upstream tarball but is present in the git
tree. It is only needed to generate the tarball itself
(together with .tarball-version file), but for some reason
upstream requires this file during build too. Just remove
call to this script for now.